+++ knock.php
@@ -7,7 +7,12 @@ class KnockCore {
 
   // Login with post variables
   public function login() {
-    $success = $this->loginUser();
+    $success = null;
+    if($this->isAuthorized()) {
+      usleep($this->o('delay') * 1000);
+      $userkey = $this->o('post.username.key');
+      $success = $this->loginUser($_POST[$userkey]);
+    }
 
     if($this->o('callback.login')) {
       return $this->o('callback.login')($success);
@@ -29,18 +34,22 @@ class KnockCore {
 
   // Check if user is authorized with post variables
   public function isAuthorized() {
-    if(!isset($_POST['username'])) return;
-    if(!isset($_POST['password'])) return;
+    $userkey = $this->o('post.username.key');
+    $passkey = $this->o('post.password.key');
 
-    $user_filepath = $this->o('path.users') . $_POST['username'] . '.php';
+    if(!isset($_POST[$userkey])) return;
+    if(!isset($_POST[$passkey])) return;
+
+    $user_filepath = $this->o('path.users') . $_POST[$userkey] . '.php';
     if(!file_exists($user_filepath)) return;
     
     $password = include($user_filepath);
-    $password_post = hash('sha256', $_POST['password']);
+    $password_post = hash('sha256', $_POST[$passkey]);
 
     if($password == $password_post) return true;
   }
 
+  // Check if user is logged in with cookie
   public function isLoggedIn() {
     $prefix = $this->o('cookie.prefix');
     $userkey = $this->o('cookie.username.key');
@@ -58,6 +67,29 @@ class KnockCore {
     $hash_cookie = hash('sha256', $hash_cookie . $this->o('salt'));
 
     if($hash == $hash_cookie) return true;
+  }
+
+  // Get the cookies expire timestamp
+  public function getCookieExpires() {
+    return $_COOKIE[$this->o('cookie.prefix')][$this->o('cookie.expires.key')];
+  }
+
+  // Refresh the cookies if the cookies will soon expire
+  public function keepAlive() {
+    $minutes = round(($this->getCookieExpires()-time())/60);
+    $diff = $minutes - $this->o('cookie.refresh');
+    if($diff < 0) {
+      return $this->refresh();
+    }
+    return true;
+  }
+
+  // Refresh the cookies, creates new hash and expire timestamp
+  public function refresh() {
+    if($this->isLoggedIn()) {
+      return $this->loginUser($_COOKIE[$this->o('cookie.prefix')][$this->o('cookie.username.key')]);
+    }
+    return true;
   }
 
   // Set options
@@ -80,28 +112,29 @@ class KnockCore {
   private function defaults() {
     return [
       'cookie.domain' => '',
-      'cookie.expires' => 2147483647,
+      'cookie.expires' => strtotime('+2 days'),
+      'cookie.expires.key' => 'expires',
       'cookie.hash.key' => 'hash',
       'cookie.path' => '/',
       'cookie.prefix' => 'knock',
+      'cookie.refresh' => 115,
       'cookie.secure' => true,
       'cookie.username.key' => 'username',
       'delay' => rand(1000, 2000),
       'path.users' => __DIR__ . '/users/',
       'path.temp' => __DIR__ . '/temp/',
+      'post.password.key' => 'password',
+      'post.username.key' => 'username',
       'salt' => '',
     ];
   }
 
   // Login user
-  private function loginUser() {
-    usleep($this->o('delay') * 1000);
-    if(!$this->isAuthorized()) return;
-
+  private function loginUser($username) {
     $hash = bin2hex(random_bytes(16));
 
-    if(!$this->writeCookie($hash)) return;
-    if(!$this->writeFile($hash)) return;
+    if(!$this->writeCookie($hash, $username)) return;
+    if(!$this->writeFile($hash, $username)) return;
 
     return true;
   }
@@ -115,18 +148,22 @@ class KnockCore {
   private function deleteCookies() {
     if(!$this->setCookie('[' . $this->o('cookie.username.key') . ']', '', 0)) return;
     if(!$this->setCookie('[' . $this->o('cookie.hash.key') . ']', '', 0)) return;
+    if(!$this->setCookie('[' . $this->o('cookie.expires.key') . ']', '', 0)) return;
 
     return true;
   }
 
   // Write cookie on login
-  private function writeCookie($hash) {
-    if(!$this->setCookie('[' . $this->o('cookie.username.key') . ']', $_POST['username'])) return;
+  private function writeCookie($hash, $username) {
+    $userkey = $this->o('post.username.key');
+    if(!$this->setCookie('[' . $this->o('cookie.username.key') . ']', $username)) return;
     if(!$this->setCookie('[' . $this->o('cookie.hash.key') . ']', $hash)) return;
+    if(!$this->setCookie('[' . $this->o('cookie.expires.key') . ']', $this->o('cookie.expires'))) return;
 
     return true;
   }
 
+  // Set cookie
   private function setCookie($key, $value, $expires = null) {
     $expires = ($expires === null) ? $this->o('cookie.expires') : $expires;
     $domain = $this->o('cookie.domain');
@@ -136,13 +173,14 @@ class KnockCore {
   }
 
   // Write temp file to disc
-  private function writeFile($hash) {
+  private function writeFile($hash, $username) {
     if(!file_exists($this->o('path.temp'))) {
       if(!mkdir($this->o('path.temp'))) return;
     }
 
     $salt = $this->o('salt');
-    $path = $this->o('path.temp') . $_POST['username'] . '.php';
+    $userkey = $this->o('post.username.key');
+    $path = $this->o('path.temp') . $username . '.php';
 
     $hash = hash('sha256', $hash . $salt);
     $content = "<?php return '" . $hash . "';";
@@ -175,5 +213,20 @@ class knock {
   public static function isLoggedIn() {
     $core = new KnockCore();
     return $core->isLoggedIn();
+  }
+
+  public static function getCookieExpires() {
+    $core = new KnockCore();
+    return $core->getCookieExpires();
+  }
+
+  public static function refresh() {
+    $core = new KnockCore();
+    return $core->refresh();
+  }
+
+  public static function keepAlive() {
+    $core = new KnockCore();
+    return $core->keepAlive();
   }
 }