+++ tests/6.login-success-check.php
@@ -1,0 +1,10 @@
+<?php
+include __DIR__ . '/../knock.php';
+
+$folder = dirname($_SERVER['PHP_SELF']);
+$link = 'http://' . $_SERVER['HTTP_HOST'] . $folder . '/7.cookie-expires.php';
+
+if(knock::isLoggedIn()) {
+  header("Location: " . $link);
+  die;
+}